Skeleton Outline for html/css/boostrap Web App
Make sure you don't change anything in the Script files or in the folders labeled css, img, js or fonts. Those are style folders for the bootstrap and the template. Go through it, read my comments on the html pages. If you see anything you want to change, change it on the html doc like a typo etc. I have commented each respective section.

For the most part there are three sections that are placeholders for ASP programming/forms. 2 on the home page. 1 on the contact page below the UI interface for the contact. Do not touch the maps section, I think I can figure the API's to display Calgary.

I have already created two webform pages that are linked to each other using respective href targets but for the ASP we need to programatically response.redirect them etc creating event handlers on the buttons.

Any questions, hit me up on Slacker. Cheers. I